Move UpdateModelPlanBeneficiaries to Beneficiary folder
